requirements.txt
streamlit
pandas
numpy
xlsxwriter
python-dotenv
openai
